+++ client/src/Components/NoData.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Alert } from "react-bootstrap";
-import config from "./config.json";
+import config from "../config.json";
 /**
  * No Data alert.
  */
